Change in ledPin lib

+++ ESP32/Solar_Powered_Irrigation_System/code/lib/Display/led_indicator/led_indicator.cpp
@@ -3,7 +3,6 @@
 
 namespace led_indicator
 {
-
   int ledPin;
   const int pwmChannel = 0;
   const int pwmFreq = 10000;
@@ -21,7 +20,7 @@ namespace led_indicator
   unsigned long pauseStart = 0;
   const int pauseDuration = 1000;
 
-  void begin(int pin)
+  void begin(uint8_t pin)
   {
     ledPin = pin;
 
@@ -82,5 +81,4 @@ namespace led_indicator
 
     ledcWrite(pwmChannel, brightness);
   }
-
 }
